Modify ApplyFriendVC code
